+++ Lib/fontbakery/specifications/googlefonts.py
@@ -138,7 +138,7 @@ expected_check_ids = \
       , 'com.google.fonts/check/glyf_unused_data'
       , 'com.google.fonts/check/currency_chars'
       , 'com.google.fonts/check/family_naming_recommendations'
-      , 'com.google.fonts/check/072' # Font enables smart dropout control in "prep" table instructions?
+      , 'com.google.fonts/check/smart_dropout'
       , 'com.google.fonts/check/maxadvancewidth'
       , 'com.google.fonts/check/074' # Are there non-ASCII characters in ASCII-only NAME table entries?
       , 'com.google.fonts/check/points_out_of_bounds'
@@ -3534,11 +3534,11 @@ def VTT_hinted(ttFont):
 
 
 @check(
-  id = 'com.google.fonts/check/072',
+  id = 'com.google.fonts/check/smart_dropout',
   conditions = ['is_ttf',
                 'not VTT_hinted']
 )
-def com_google_fonts_check_072(ttFont):
+def com_google_fonts_check_smart_dropout(ttFont):
   """Font enables smart dropout control in "prep" table instructions?
 
   B8 01 FF    PUSHW 0x01FF